Restore into-make-service-with-connect-info
